+++ app/imports/ui/lexes/RichEditor.js
@@ -1,7 +1,42 @@
 import React from 'react';
-import { Editor, EditorState, RichUtils } from 'draft-js';
+import {
+  CompositeDecorator,
+  Editor,
+  EditorState,
+  RichUtils,
+} from 'draft-js';
 
-export class RichEditorExample extends React.Component {
+function findLinkEntities(contentBlock, callback, contentState) {
+  contentBlock.findEntityRanges(
+    (character) => {
+      const entityKey = character.getEntity();
+      return (
+        entityKey !== null &&
+        contentState.getEntity(entityKey).getType() === 'LINK'
+      );
+    },
+    callback
+  );
+}
+
+const Link = (props) => {
+  const {url} = props.contentState.getEntity(props.entityKey).getData();
+  return (
+    <a href={url} style={styles.link}>
+      {props.children}
+    </a>
+  );
+};
+
+const decorator = new CompositeDecorator([
+  {
+    strategy: findLinkEntities,
+    component: Link,
+  },
+]);
+
+export class PolylexRichEditor extends React.Component {
+
   onChange = editorState => {
     let reference = this.props.reference;
     this.props.onChange(reference, editorState);
@@ -17,36 +52,108 @@ export class RichEditorExample extends React.Component {
       this.refs.editor.focus();
     } 
   }
+  
+  constructor(props) {
+    super(props);
+    let newEditorState = EditorState.set(this.props.editorState, {decorator: decorator});
+    this.onChange(newEditorState);
+    this.state = {
+      editorState: newEditorState,
+      showURLInput: false,
+      urlValue: '',
+    };
+    this.promptForLink = this._promptForLink.bind(this);
+    this.onURLChange = (e) => this.setState({urlValue: e.target.value});
+    this.confirmLink = this._confirmLink.bind(this);
+    this.onLinkInputKeyDown = this._onLinkInputKeyDown.bind(this);
+    this.removeLink = this._removeLink.bind(this);
+  }
 
-  handleKeyCommand = command => {
-    const { editorState } = this.props;
-    
-    const newState = RichUtils.handleKeyCommand(editorState, command);
-   
-    if (newState) {
-      this.onChange(newState);
-      return true;
+  _promptForLink(e) {
+    e.preventDefault();
+    let editorState = EditorState.set(this.props.editorState, {decorator: decorator});
+    this.onChange(editorState);
+    const selection = editorState.getSelection();
+    if (!selection.isCollapsed()) {
+      const contentState = editorState.getCurrentContent();
+      const startKey = editorState.getSelection().getStartKey();
+      const startOffset = editorState.getSelection().getStartOffset();
+      const blockWithLinkAtBeginning = contentState.getBlockForKey(startKey);
+      const linkKey = blockWithLinkAtBeginning.getEntityAt(startOffset);
+
+      let url = '';
+      if (linkKey) {
+        const linkInstance = contentState.getEntity(linkKey);
+        url = linkInstance.getData().url;
+      }
+
+      this.setState({
+        showURLInput: true,
+        urlValue: url,
+      }, () => {
+        setTimeout(() => this.refs.url.focus(), 0);
+      });
     }
-    return false;
-  };
+  }
 
-  onTab = e => {
-    const maxDepth = 4;
-    this.onChange(RichUtils.onTab(e, this.props.editorState, maxDepth));
-  };
+  _confirmLink(e) {
+    e.preventDefault();
+    const {urlValue} = this.state;
+    let editorState = EditorState.set(this.props.editorState, {decorator: decorator});
+    this.onChange(editorState);
+    const contentState = editorState.getCurrentContent();
+    const contentStateWithEntity = contentState.createEntity(
+      'LINK',
+      'MUTABLE',
+      {url: urlValue}
+    );
+    const entityKey = contentStateWithEntity.getLastCreatedEntityKey();
+    const newEditorState = EditorState.set(editorState, { currentContent: contentStateWithEntity });
+    const newES = RichUtils.toggleLink(
+      newEditorState,
+      newEditorState.getSelection(),
+      entityKey
+    );
+    this.onChange(newES);
+    this.setState({
+      editorState: newES,
+      showURLInput: false,
+      urlValue: '',
+    }, () => {
+      setTimeout(() => this.focus(), 0);
+    });
+  }
+
+  _onLinkInputKeyDown(e) {
+    if (e.which === 13) {
+      this._confirmLink(e);
+    }
+  }
+
+  _removeLink(e) {
+    e.preventDefault();
+    let editorState = EditorState.set(this.props.editorState, {decorator: decorator});
+    this.onChange(editorState);
+    const selection = editorState.getSelection();
+    if (!selection.isCollapsed()) {      
+      this.onChange(RichUtils.toggleLink(editorState, selection, null));
+    }
+  }
 
   toggleBlockType = blockType => {
     this.onChange(RichUtils.toggleBlockType(this.props.editorState, blockType));
   };
+
   toggleInlineStyle = inlineStyle => {
     this.onChange(
       RichUtils.toggleInlineStyle(this.props.editorState, inlineStyle)
     );
   };
-  
-  render() {
-    const { editorState, reference } = this.props;
 
+  render() {
+    let editorState = EditorState.set(this.props.editorState, {decorator: decorator});
+    const { reference } = this.props;
+    
     // If the user changes block type before entering any text, we can
     // either style the placeholder or hide it. Let's just hide it now.
     let className = 'RichEditor-editor';
@@ -57,35 +164,66 @@ export class RichEditorExample extends React.Component {
         className += ' RichEditor-hidePlaceholder';
       }
     }
-    
-    return (
-      <div className="RichEditor-root">
-        <BlockStyleControls
-          editorState={editorState}
-          onToggle={this.toggleBlockType}
-        />
-        <InlineStyleControls
-          editorState={editorState}
-          onToggle={this.toggleInlineStyle}
-        />
-        <div className={className} onClick={this.focus}>
-          <Editor
-            blockStyleFn={getBlockStyle}
-            customStyleMap={styleMap}
-            editorState={editorState}
-            handleKeyCommand={this.handleKeyCommand}
-            onChange={this.onChange}
-            onTab={this.onTab}
-            placeholder=""
-            ref={reference}
-            spellCheck={true} 
+
+    let urlInput;
+    if (this.state.showURLInput) {
+      urlInput =
+        <div style={styles.urlInputContainer}>
+          <input
+            onChange={this.onURLChange}
+            ref="url"
+            style={styles.urlInput}
+            type="text"
+            value={this.state.urlValue}
+            onKeyDown={this.onLinkInputKeyDown}
           />
+          <button onMouseDown={this.confirmLink}>
+            Valider
+          </button>
+        </div>;
+    }
+
+    return (      
+      <div className="RichEditor-root">
+        <div style={styles.root}>
+          <div style={styles.buttons}>
+            <button
+              type="button"
+              onMouseDown={this.promptForLink}
+              style={{marginRight: 10}}>
+              Ajouter un lien
+            </button>
+            <button type="button" onMouseDown={this.removeLink}>
+              Supprimer un lien
+            </button>
+          </div>
+          {urlInput}
+          <BlockStyleControls
+            editorState={editorState}
+            onToggle={this.toggleBlockType}
+          />
+          <InlineStyleControls
+            editorState={editorState}
+            onToggle={this.toggleInlineStyle}
+          />
+          <div className={className} style={styles.editor} onClick={this.focus}>
+            <Editor
+              blockStyleFn={getBlockStyle}
+              customStyleMap={styleMap}
+              editorState={editorState}
+              handleKeyCommand={this.handleKeyCommand}
+              onChange={this.onChange}
+              onTab={this.onTab}
+              placeholder=""
+              ref={reference}
+              spellCheck={true} 
+            />
+          </div>
         </div>
       </div>
     );
   }
 }
-
 
 // Custom overrides for "code" style.
 const styleMap = {
@@ -96,6 +234,7 @@ const styleMap = {
     padding: 2,
   },
 };
+
 function getBlockStyle(block) {
   switch (block.getType()) {
     case 'blockquote':
@@ -125,6 +264,7 @@ class StyleButton extends React.Component {
     );
   }
 }
+
 const BLOCK_TYPES = [
   //{ label: 'H1', style: 'header-one' },
   //{ label: 'H2', style: 'header-two' },
@@ -137,6 +277,7 @@ const BLOCK_TYPES = [
   { label: 'OL', style: 'ordered-list-item' },
   //{ label: 'Code Block', style: 'code-block' },
 ];
+
 const BlockStyleControls = props => {
   const { editorState } = props;
   const selection = editorState.getSelection();
@@ -158,12 +299,14 @@ const BlockStyleControls = props => {
     </div>
   );
 };
+
 var INLINE_STYLES = [
   { label: 'Bold', style: 'BOLD' },
   { label: 'Italic', style: 'ITALIC' },
 //  { label: 'Underline', style: 'UNDERLINE' },
 //  { label: 'Monospace', style: 'CODE' },
 ];
+
 const InlineStyleControls = props => {
   var currentStyle = props.editorState.getCurrentInlineStyle();
   return (
@@ -179,4 +322,37 @@ const InlineStyleControls = props => {
       )}
     </div>
   );
+};
+
+const styles = {
+  root: {
+    fontFamily: '\'Georgia\', serif',
+    padding: 20,
+    width: 600,
+  },
+  buttons: {
+    marginBottom: 10,
+  },
+  urlInputContainer: {
+    marginBottom: 10,
+  },
+  urlInput: {
+    fontFamily: '\'Georgia\', serif',
+    marginRight: 10,
+    padding: 3,
+  },
+  editor: {
+    border: '1px solid #ccc',
+    cursor: 'text',
+    minHeight: 80,
+    padding: 10,
+  },
+  button: {
+    marginTop: 10,
+    textAlign: 'center',
+  },
+  link: {
+    color: '#3b5998',
+    textDecoration: 'underline',
+  },
 };
